Introduce `Request` class instead of `Input` class
In order to clarify the subject of class name
