+++ cl/bs.php
@@ -19,7 +19,8 @@ require CL_CLASSES . 'AutoLoader.php';
     'DB'        => '\CodeLapse\DB',
     'File'      => '\CodeLapse\File',
     'Form'      => '\CodeLapse\Form',
-    'Input'     => '\CodeLapse\Input',
+    'Req'       => '\CodeLapse\Request',
+    /* deprecated */ 'Input'     => '\CodeLapse\Input',
     'Mail'      => '\CodeLapse\Mail',
     'Pager'     => '\CodeLapse\Pager',
     'Security'  => '\CodeLapse\Security',
